Create an Operator object to handle the random choice of the operator
This came along with a refactor of the code
